fix: strip private-unviewable lists from getEventsForList results
getEventsForList hydrated event.lists with every list the event belonged
to, including private lists the viewer couldn't access. SavedByModal on
the list detail page explicitly trusts the server's filtering (per the
contract established by feeds.ts queryFeed/queryGroupedFeed), so private
list names/slugs were leaking on any list detail page that rendered
events belonging to multiple lists.

Fix: extract getViewableListIds from feeds.ts into model/lists.ts (shared
source of truth), then apply it in getEventsForList to strip unviewable
lists from each event.lists before returning. feeds.ts now imports the
same helper so behavior cannot drift.

Addresses review feedback: every code path that populates event.lists
must enforce viewer visibility — client-side filtering has been
explicitly removed in favor of server-side trust.
